Refactor Outfeeds in prep for insgle image compilation and optimise performance
Summary:
We use the spsc queue with pre allocated memory for communication between
the outfeed connect stream thread and the new outfeed io thread.
This queue means that all the callback has to do is a memcpy.

Introduced Front and Back to the queue rather than push/pop to write the memory
already allocated.

We then create an output queue which is serialized when the CPU op tries to dequeue.
The outfeeds do have a lock however this only prevents reads from the queue during the
enging execution and not before, but at least now we throw an error (T8971).

This has mainly been done for the next patch where we will suport single image compilation.
This also optimizes the outfeed performance as the outfeed callback does not do a memcpy.

This also allows us to have multiple outfeed queues as we no longer use the transfer manager (but this has not been enabled in this diff).
This also allows us to not have to dequeue an outfeed immedietly as the transfer manager would reset the outfeed as soon as a different XLA
graph was executed.

Fixes T7218

Test Plan: All tests pass

Reviewers: davidn, jakeh, stephenm, romand

Reviewed By: jakeh

Maniphest Tasks: T7218

Differential Revision: https://phabricator.sourcevertex.net/D12319