Add IPUEstimator with train API
Summary:
Start working on an IPUEstimator by first implementing the train API.

Checkpointing and warm-starting is supported.

Augment the user-provided model function with IPU support and
call the train function of the Estimator base class which
uses this augmented model function. This approach is similar
to one taken by the TPUEstimator.

Note that the Estimator base class has started disallowing
overriding its member functions recently, which we work
around by replacing the _assert_members_are_not_overridden
function. See https://github.com/tensorflow/tensorflow/issues/18314

Use a hook to increment the global_step variable on the CPU after
each session run. If we were to increase it on the IPU, a separate
engine would be compiled and swapped in just for increasing the counter.
This same hook will also request stopping the training after reaching
the desired number of steps (it must be done in the same hook because
hook execution order is undefined).

Co-authored-by: Shaurya Sharma <shauryas@graphcore.ai>

Test Plan: Added unit tests. Trained a simple CNN MNIST model.

Reviewers: georgep, davidn

Reviewed By: georgep, davidn

Subscribers: shauryas

Differential Revision: https://phabricator.sourcevertex.net/D12450