IPUEstimator: Add support for evaluation and prediction
Summary:
Implement the `evaluate()` and `predict()` member functions on `IPUEstimator`.
These use the same `iterations_per_loop` configuration as `train()` and compile
an evaluation or prediction loop for the IPU. The output is fed back using
an `IPUOutfeedQueue`, with `IPUOutfeedMode.LAST` for `evaluate()` (the last
value of the accumulated evaluation metrics are returned) and
`IPUOutfeedMode.ALL` for `predict()`. The extra dimension of the outfed tensor
caused by the loop is flattened before yielding the results from `predict()`,
which makes behaviour equivalent to `tf.estimator.Estimator`, such that it also
supports the `yield_single_examples` parameter.

Test Plan:
New unit tests and compared CIFAR-10 model output with CPU results

Co-authored-by: Shaurya Sharma <shauryas@graphcore.ai>

Reviewers: georgep, davidn

Reviewed By: davidn

Subscribers: shauryas

Differential Revision: https://phabricator.sourcevertex.net/D12717