Remove check for AllReduce operations in replicated graph
Summary:
Previously we required there to be an `AllReduce` operation
in a graph that was replicated and had an infeed. This would
typically not be the case when doing replicated prediction
where the replicas operate independently on examples from
the same infeed. This check could be overridden with the
environment variable `TF_POPLAR_FLAGS=--force_replicated_mode`.

The plan after removing this is to take a more targeted
approach, where in the `IPUEstimator` we can enforce
that the training graph contains an cross replica sum,
while the prediction graph does not have a such a
requirement. This is similar to what the `TPUEstimator`
does. Of course, if the user forgets the `CrossReplicaOptimizer`
when using the `Session` API directly, it will now go unnoticed.

Test Plan:
Repurposed now obselete test to do a replicated
identity function. Not really sure how useful it is...

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Subscribers: georgep

Differential Revision: https://phabricator.sourcevertex.net/D12885