Fix loop body input arity check
Summary:
Previously there was a mismatch between the number of
input arguments that were assumed by the check in
xla.check_function_argument_count and the number of
arguments actually used when calling the loop body.

This was because we have some logic where we unpack
a dict (when calling) if it is the only element
given by the dequeue operation.

This patch starts using the same logic also when
doing the check to avoid this mismatch.

Test Plan: Unit tests

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Subscribers: georgep

Differential Revision: https://phabricator.sourcevertex.net/D12931