Add extended custom op example to docs
Summary:
Adding a new section to the docs on custom ops, with an extended example, included in the unit tests.

Also extended the DSO file symbols to include an IsElementwise and an AllocateInput.

Have not wired up the AllocteInput yet.

Test Plan: all tests ok, new one added

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Subscribers: georgep

Tags: #tensorflow

Differential Revision: https://phabricator.sourcevertex.net/D13245