Don't register support for the DT_RESOURCE type
Summary:
TF2 checks that the _Arg op does't have conflicting types, and
registers the _Arg op with a DT_RESOURCE (as does TF1).  So we
don't need to register this type at all.

Note also that the CPU and GPU backends do not reigster support
for the resource type.

Test Plan: All tests ok

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Subscribers: georgep

Differential Revision: https://phabricator.sourcevertex.net/D13435