Supress warning when rebalancing wouldn't help.
Summary:
In the case of a very dense lookup, the output of the first stage of a two-stage lookup would be larger than the original embedding.
In this case it's better to fallback to the TF embedding_lookup. This also means we supress the warning, so users don't try to reshape their embedding.

Test Plan: All CI tests pasxs

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Differential Revision: https://phabricator.sourcevertex.net/D13663