Make HloInstruction::RemoveUser amortized constant time
Summary:
Cherry picked from TF2 to make users() order stable between TF1/2.
Change to multi_ipu_test removes the compute set difference between TF1/2, their source code is the same now.
Ref T28597. TF1.15 Only.

Previously it was linear in the number of users because we had to traverse
an std::vector to find the element to remove. After this change we store
the index inside the vector in the map used for fast lookup instead.

PiperOrigin-RevId: 266884015

Test Plan: All tests pass.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T28597

Differential Revision: https://phabricator.sourcevertex.net/D35509