Lower constants for pipelines in TF2XLA and XLA
Summary:
We need to resolve and propagate constants so that operations in XLA
which takes constant runtime tensors as inputs still work (such as
transpose).

This means we might end up with empty stages and/or modifications to
parameters which are not associated to a gradient (for example constant
gradients).

To prevent this, we insert stateful no-ops into pipeline stages to make
sure they are not deleted by DCE before we lower operations into them.
We also go through all the users of parameters and do any nessecarry
lowering.

Fixes T11044

Test Plan: all tests pass, added new tests.

Reviewers: #tensorflow, jakeh

Reviewed By: #tensorflow, jakeh

Maniphest Tasks: T11044

Differential Revision: https://phabricator.sourcevertex.net/D14274