Add IPUMultiWorkerStrategy
Summary:
This is a distribution strategy for synchronous training using IPUs
on multiple workers with between-graph replication.

It places variables on the host device of each worker, and does
multi-worker all-reduce to to keep the variables in sync, using
TensorFlow's implementation of collective operations over gRPC.

It is the responsibility of the user to place the operations on the
IPU, while this strategy will make sure that the variables are kept
on the host and in sync between the multiple workers.

When used during training with an `Optimizer`, this means that the
variables will be streamed from the host to the IPU when needed,
and that the gradients will be streamed back to the host and then
all-reduced across the workers. Then the workers will do identical
updates to their copies of the variables. In other words,
`optimizer.compute_gradients()` is done on the device, while
`optimizer.apply_gradients()` is done on the host. All the "slot"
variables used by the optimizer (e.g. the accumulator in momentum)
are kept only in host memory and never used on the device, saving
device memory.

It does not yet work with the `IPUEstimator` which tries to compile
one big XLA cluster for everything, which conflicts with how this
strategy will split out the `optimizer.apply_gradients()` part and
place it on the host.

Test Plan: Added new tests

Reviewers: #tensorflow, georgep, davidn

Reviewed By: #tensorflow, georgep, davidn

Subscribers: shauryas

Differential Revision: https://phabricator.sourcevertex.net/D14508