Zero feed counters at the start of every execution
Summary:
The counters for the feed buffers should be reset to zero at the
start of each execution, rather than being left with the values that
they previously contained.

This has the added benefit that their liveness does not wrap around
through the top of the program loop, so that they don't have to be
present during weight upload/download.

Test Plan: All tests ok

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Differential Revision: https://phabricator.sourcevertex.net/D14542