ipu_estimator: Add IPUEstimatorSpec with host_call parameter
Summary:
This can be used to take output from the model_fn compiled for
the IPU and pass it to a function that is placed on the host.
This is suitable for distributed training where e.g. the gradients
must be passed to the host for global reduction.

This host_call parameter is similar to the one in TPUEstimatorSpec.

Right now the only reason to use IPUEstimatorSpec instead of
EstimatorSpec is to supply this host_call parameter. We might
be more strict and require IPUEstimatorSpec in the future.

Remove use_xla_auto_clustering flag which was very unreliable, now
the model_fn is always compiled explicitly, while the host_call is not.

It makes this aspect of the distribution strategy a somewhat leaky
abstraction in the IPUEstimator, but it is very explicit and reliable
compared to other alternatives.

Reviewers: #tensorflow, shauryas, davidn

Reviewed By: #tensorflow, davidn

Subscribers: davidn

Maniphest Tasks: T11312

Differential Revision: https://phabricator.sourcevertex.net/D14828