ipu_multi_worker_strategy: Support sync-on-read variables
Summary:
Use custom types for distributed variables.

This lets us use sync-on-read variables properly, which
fixes usage of the batch normalization layer. We add our
own types to make it more explicit, and we might want to
override some behaviour later as well.

The idea with sync-on-read variables is that writes dominate
over reads, so it makes sense to place the per-replica
variables on the IPU of the corresponding workers and
only do the transfer and reduction on CPU when they
are read (when the cross-replica value is needed).

A typical example of this is the moving mean and variance
in a batch normalization layer that is calculated during
training, but only used (e.g. read) later when saving the
model or doing inference.

Reviewers: #tensorflow, shauryas

Reviewed By: shauryas

Maniphest Tasks: T11643

Differential Revision: https://phabricator.sourcevertex.net/D15244