Add custom ops for sending/receiving tensors from IPU to CPU
Summary:
Add IpuSendToHostOp and IpuRecvAtHostOp that can be used to
send a tensor from the IPU (inside the XLA cluster) and
receive it at the host side.

The IpuSendToHostOp is lowered to the XLA SendToHost op,
which again is lowered to a pair of Send and SendDone ops.
SendDone is documented to block until the send is done,
so this is where we lower to a synchronous copy into
a Poplar stream (also this is where we can most easily
attach the frontend attribute with the rendezvous key).

The IpuRecvAtHostOp is an async op that completes when
the tensor is produced via the rendezvous mechanism
from the Poplar executor.

Serialization of engines using send ops is not yet
implemented.

Test Plan: Added some new tests.

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Subscribers: georgep

Differential Revision: https://phabricator.sourcevertex.net/D16081