ipu_estimator: Improve handling of datasets of various shapes
Summary:
Our handling of datasets of various shapes were a bit different
compared to the regular Estimator. For example when passing a
"flat" dataset which contains a dict on the top level (without
nesting in a tuple), everything was passed as features with labels
as None in the Estimator, while the IPUEstimator gave an error
(this is undocumented in Estimator).

Make the IPUEstimator dataset handling as similar as possible to
that of the Estimator, and use parameterized tests to make sure
that the relevant tests behave the same with the IPUEstimator
and the Estimator.

Based on customer feedback through Mark Fowden.

Reviewers: #tensorflow, davidn

Reviewed By: #tensorflow, davidn

Maniphest Tasks: T13261

Differential Revision: https://phabricator.sourcevertex.net/D17522