inplace_util: Mark Send as inplace read-only
Summary:
The Send instruction must be marked as inplace since the data that
it is supposed to send is passed as a reference through to the
SendDone instruction (that actually sends it in our backend), so
it must not be modified in between.

Previously we could observe that e.g. an inplace Add instruction
could be scheduled in between the Send and SendDone, resulting in
sending the wrong value.

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Subscribers: georgep

Differential Revision: https://phabricator.sourcevertex.net/D17598