tf2xla: Add dependencies in a deterministic order
Summary:
Similar to the traversal of the nodes in alphabetic order
(as a tie-breaker in the topological order), traverse
the input edges in an deterministic order by the source
node ID. This makes sure that the XLA HLO is deterministic
and that we avoid false misses in the executable cache.

This is needed to get the control dependencies added in
a deterministic order, and can be reverted when we remove
support for control dependencies.

Have not found a small test case that reproduces the cache miss
issue, but can be reproduced with the ipu_estimator_example.py.

See T13186.

Reviewers: #tensorflow, davidn

Reviewed By: #tensorflow, davidn

Subscribers: stephenm, davidn

Differential Revision: https://phabricator.sourcevertex.net/D17895