poplar_executable: Make device hash computation stateless
Summary:
The computation of the device hash used the previous device
hash as input (as the starting point). This caused false
executable cache misses when configuring the same device
multiple times. Even though the configuration was exactly
the same, the computed hash was different because of this
statefulness.

This can be tested by using multiple TF Graphs in the
same process, since each graph will have its own XLA
compilation cache, so the last level cache is the
Poplar executable cache.

This is the first step towards re-using executables
between successive calls to "ipu_estimator.train()".
The next step is to clean up the infeed and re-use
the name between calls.

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Differential Revision: https://phabricator.sourcevertex.net/D17992