ipu_infeed_queue: Add deleter
Summary:
Add a delete operation that can be used to delete
the dataset iterator. This allows re-using the
name of the infeed, which again allows re-using
cached executables if everything else is also the same.

Change the destruction order of the members in
InfeedDatasetIterator since they borrow some pointers
between each other and have specific lifetime
requirements that became apparent when destroying them
in the wrong order (illegal memory accesses in the
destructors).

Rename IPUConsumeDataset to IPUCreateDatasetIterator
which is slightly more descriptive (IMO) and symmetric
with the new IPUDeleteDatasetIterator.

The `deleter` name is not entirely symmetric with the
`initializer` name, but it is more clear in not
encouraging re-use of the infeed. Suggestions welcome.

Drive-by fix of includes according to linter.

Reviewers: #tensorflow, shauryas, davidn

Reviewed By: #tensorflow, davidn

Subscribers: davidn

Differential Revision: https://phabricator.sourcevertex.net/D18045