ipu_estimator: Re-use infeed IDs to get more executable caching
Summary:
A common use case for an `Estimator` is to alternate between training
and evaluation. With the `IPUEstimator` this would cause recompilation
for every alternation because the infeed ID would be incremented every
time, causing a miss in the executable cache. This patch avoids
recompiling the training engine.

Make the `IPUEstimator` delete the infeed, such that the same ID
can be re-used and we get a cache hit (as long as the `input_fn` is
compatible, e.g. gives the same data shape) for the train function.

The evaluation engine, however, is still recompiled every time
because it uses outfeeds which we cannot delete yet.

Use a `SessionRunHook.end` callback to delete the infeed when done.
Deletion in `__del__` might be necessary to ensure that the infeed
is actually deleted in all circumstances, e.g. when an exception
is thrown or when the generator returned by `predict()` is deleted.

Also take care of the case where multiple `IPUEstimators` are used
from different threads, thus needing unique IDs after all. (This
strange usage happens in the `IPUMultiWorkerMirroredStrategy` test.)

See T13516.

Reviewers: #tensorflow, shauryas, georgep

Reviewed By: #tensorflow, georgep

Differential Revision: https://phabricator.sourcevertex.net/D18069