Add IPUPipelineEstimator
Summary:
The `IPUPipelineEstimator` is similar to `IPUEstimator`, but
with support for pipelined execution across multiple IPUs.
This is based on our existing pipelining ops.

The reason for making a new class is that the difference
is quite substantial both in API and in features such that
it seems much cleaner to provide two different APIs that
more clearly communicate what is supported in each of them.

The major API difference from the IPUEstimator is that the
provided `model_fn` must return an `IPUPipelineEstimatorSpec`
that contains the information needed for pipelined execution
(and that omits what is not supported from `IPUEstimatorSpec`).

The IPUEstimator is refactored to enable code reuse, but
should have no functional changes.

Added new tests, documentation and example.

The example is very similar to the existing CIFAR-10 example
for `IPUEstimator`, only ported to the `IPUPipelineEstimator`.

Depends on D18961.

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Subscribers: georgep, davidn

Maniphest Tasks: T13179

Differential Revision: https://phabricator.sourcevertex.net/D18962