ipu_estimator: Cache the global_step variable on the IPU
Summary:
Before this patch reading the global_step in the model function on
the IPU did not work because it introduces a reference connection
from the CPU to the IPU, causing colocation_graph.cc to force
the cluster to the CPU with an error like this (even with
allow_soft_placement=False, which is IMO surprising):

tensorflow/core/common_runtime/colocation_graph.cc:244] Ignoring
device specification /device:IPU:0 for node 'cluster' because the
input edge from 'global_step' is a reference connection and already
has a device field set to /device:CPU:0

However, we are allowed to cache the variable on the IPU such
that the model function can read it without a reference connection.
It is however still not allowed to write to it.

This is useful for e.g. a learning rate schedule based on the
number of training steps.

Reviewers: #tensorflow, shauryas, georgep

Reviewed By: #tensorflow, georgep

Subscribers: markf, mariok

Differential Revision: https://phabricator.sourcevertex.net/D18540