ipu_estimator: Avoid relying on unspecified metric tensors
Summary:
The metric functions in the Python API return a tuple of
`(value_tensor, update_op)`. Previously we enqueued the `value_tensor`
on the outfeed as the metric results. However it turns out we cannot
depend on the `value_tensor` as it is unspecified whether it
is evaluated before or after the `update_op`. For example there are no
control dependencies between the `assign_add()` update ops and the tensor
in `metrics_impl.mean()`. There does however seem to be a guarantee that
the `update_op` returns the updated value (see e.g.
MeanTest.testUpdateOpsReturnsCurrentValue), so we will just ignore the
`value_tensor` and use the result of the `update_op` instead.

Test Plan: All tests now pass in poplar/r2.1/merge

Reviewers: #tensorflow, georgep, davidn

Reviewed By: #tensorflow, georgep, davidn

Differential Revision: https://phabricator.sourcevertex.net/D18660