Make sure the ConvNormMatcher is matching a convolution
Summary:
The ConvNormMatcher looks for an instruction whos output feeds two later instructions (batch-norm and batch-grad).
This instruction is currently assumed to always be a convolution, but that is not always true.
e.g.
In some cases the gamma param might be converted (cast) f32->f16 and its result feeds the same two instructions (batch-norm and batch-grad).
The ConvNormMatcher will match the convert. This leads to bogus attempts to inject a cloned convolution.

This change makes sure the instruction is a convolution before accepting the match.

NOTE:
This change makes ConvNormMatcher consistent with ConvNormLMMatcher.

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Subscribers: georgep

Maniphest Tasks: T13659

Differential Revision: https://phabricator.sourcevertex.net/D19093