Add option to only extract the events types.
Summary:
In Eager mode we end up with several compile events which are hard to combine, but in some tests we only want to know if something ran on the CPU or the IPU.
This patch adds an intermediate state in the parsing where we only return the unprocessed list of events and their types. (Enough to determine if the IPU was targeted/used).

Test Plan: All tests pass

Reviewers: #tensorflow, davidn

Reviewed By: #tensorflow, davidn

Differential Revision: https://phabricator.sourcevertex.net/D19433