reduce_scatter: New custom popops op
Summary:
The idea is that this can be used to optimize the hierarchical
reductions for distributed training.

Then we can do reduce_scatter() between the replicas over the IPU
links, send the result from each replica to the host, and concatenate
(gather) the results there to form the per-worker reduced result. This
is then passed to an inter-worker allreduce routine to get the final
global result.

This has some benefits compared to doing cross_replica_sum()
and sending the result from one of the IPUs to the host (which
is what we currently do):

* reduce_scatter() is less work than cross_replica_sum()
* Each replica only needs to transfer ceil(num_elements / num_replicas)
  elements to the host (and they can be done in parallel)

Fixes T14756.

Test Plan: New replicated test. Existing tests pass.

Reviewers: #tensorflow, jakeh

Reviewed By: #tensorflow, jakeh

Subscribers: jakeh

Maniphest Tasks: T14756

Differential Revision: https://phabricator.sourcevertex.net/D19772