Allow unbounded work queue to be spawned on a specific NUMA node
Summary:
https://github.com/tensorflow/tensorflow/pull/35640

This patch allows us to create unbounded thread pools pinned to a specific node. This is useful when we want to pin workloads to a specific CPU node (for example NUMA aware dataset iterators, which are similar to MultiDeviceIterator where there is an iterator per NUMA node).

This should not break the public release but I assume the non-default internal implementation will need updating.

Reviewers: #tensorflow, hakons

Reviewed By: #tensorflow, hakons

Differential Revision: https://phabricator.sourcevertex.net/D19891