pipelining_ops: Make gradient shape inference work
Summary:
In the _pipeline_stage_grad() function, func_graph.output_shapes is
set by the function_def_to_graph() function, which seems to need
some help from the "_output_shapes" attribute to be able to set
the shapes for the functions where op_needs_rewrite is True.
Otherwise they are unknown, and then the gradients also end up
with unknown shapes.

I do not understand exactly where/if this is supposed to be set
automatically, but adding it explicitly here seems OK for now.

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Differential Revision: https://phabricator.sourcevertex.net/D20143