ipu_compiler: Fix exception context
Summary:
Using PEP 409 -- Suppressing exception context.

Before this, if an exception was thrown from the function being compiled,
the attempt at detecting the IPUEstimatorSpec would augment the
exception context like this:

    Traceback (most recent call last):
      File "ipu_compiler.py", line 107, in compile
        outputs = inspect.trace()[-1][0].f_locals['outputs']
    KeyError: 'outputs'

    During handling of the above exception, another exception occurred:

    *Here comes the real exception*

After this fix, the first part is discarded such that the
user only sees the real exception.

Test Plan: Did not find an easy way to test this, only manunal testing

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Differential Revision: https://phabricator.sourcevertex.net/D20207