ipu_multi_worker_strategy_test: Improve pipeline testing
Summary:
Inheriting directly from googletest.TestCase avoids some
setup code that will seed the random number generator
random::New64() from /dev/urandom before we fork off the
child processes.

The problem with the children having the same seed is that
the GetUniqueRequestId() will return the same for both
workers which results in a collision ("the same gRPC
request was received twice"). And we cannot easily
re-seed since it is a statically initialized variable in C++.

This fix allows us to do more thorough pipeline testing.

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Differential Revision: https://phabricator.sourcevertex.net/D20385