outside_compilation_scope: Split XlaRecvAtHost nodes to get overlap
Summary:
Split XlaRecvAtHost nodes with multiple outputs such that the new nodes
have only one output each. This rewrite avoids the synchronization point
of a single node that waits for all the necessary data to satisfy all of
its outputs. The split nodes can complete when only their data is ready.

For example, when doing an allreduce of all the gradients in an
outside_compilation_scope, we can now start allreducing the first
gradient received on the host without waiting for all the others.

Then we can also remove the OutputAccumulator which acted as this
synchronization point.

Test Plan: All existing tests pass. Inspecting overlap in execution trace.

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Differential Revision: https://phabricator.sourcevertex.net/D20442