Switch to a subclass of Function returned by ipu.function
Summary:
In order to be able to operate the advanced API methods on the
result of annotating a function with an ipu.function, we need
to return a subclass of the eager Function object.

This change re-creates the tf.function annotation wrapper call,
and returns a subclass of the Function.

I have also added a test to verify that the ipu.function can be
used to annotate a class member.

Fixes: T15353

Test Plan: All tests pass

Reviewers: #tensorflow, georgep, anthonyb

Reviewed By: #tensorflow, georgep, anthonyb

Subscribers: anthonyb

Differential Revision: https://phabricator.sourcevertex.net/D20453