ipu_multi_worker_strategy_test: Test with IPUPipelineEstimator
Summary:
Add a test for the combination of IPUMultiWorkerStrategy and IPUPipelineEstimator.

Fix issue discovered with duplicated op names that was caused by the
allreduce op inside a FunctionDef (for allreducing the gradients in
the weight update function) having the same name as the allreduce op
in the outer GraphDef (for allreducing the loss). When the
extract_outside_compilation_pass extracted the op from the function
it collided with the op in the outer graph.

Reduce the risk of this happening by using a name scope with the
outside compilation scope.

Reviewers: #tensorflow, davidn

Reviewed By: #tensorflow, davidn

Subscribers: davidn

Maniphest Tasks: T14530

Differential Revision: https://phabricator.sourcevertex.net/D20637