ipu_multi_worker_strategy: Place variables and ops on IPU by default
Summary:
Change/break the API while we still have the opportunity (it is only
in "preview" mode) :-)

Now we will place the variables and ops on the IPU by default. Placing
variables on the IPU is the approach that gives the least amount of
changes in user code since the `optimizer.apply_gradients()` will
automatically use an outside compilation scope to do the gradient
allreduce, by implicitly streaming the data to and from the host.

When placing variables on the host, the user must also place the
`optimizer.apply_gradients()` call on the host explicitly (either
by explicit device placement outside an XLA compilation or by using
the `host_call` feature of the IPUEstimator). This might require
a bit more changes in user code, but allows for placing momentums
etc. in host memory to save IPU memory.

Fix up the IPUEstimator model_fn that now might be called with
an IPU device scope active, so we need to explicitly override
this with a host device scope when needed.

Elaborate a bit better in the docs.

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Maniphest Tasks: T14820

Differential Revision: https://phabricator.sourcevertex.net/D20820