Defer allocation of inputs until we have an allocation target or the tensor is being used
Summary:
An Hlo instruction visitor which deferres allocation of tensors until we can
either allocate them or we are forced to use them.

In order to postpone the allocation, handlers for the inputs (Parameters and
Infeeds) check whether they have an allocation target for that input tensor
and if they do not, they use DeferredAllocations structure to indicate that a
particular input is deferred. When an input is deferred, we provide:
- the location of the input
- the allocation function which is called to allocate the tensor in the
Poplar graph.
- the post processing function which can be used to add any stream copies to
a sequence.
These functions will be called when an allocation is being made.

We also specialize few operations which do not need the tensor be alloacted
as they do not perform any calculations.

Specialization of the handle for GetTupleElement instruction which checks if
any of its inputs are deferred, and if they are:
- and we have an allocation target for this location (usually generated by
the ForwardAllocation pass) then we allocate the tensor and propagate the
allocation information.
- otherwise if we do not have an allocation target, we also marks this
location as deferred.

If the input to the GTE already has a tensor then we just forward the tensor.
Specialization of the handle for Tuple instructions, similarly checks all the
input locations, and if any of them are deferred inputs we mark the output as
deferred too, otherwise we forward the tensor.

For instructions with callsites - While loops, Repeat loops, Pipelines and
Pipeline Stages, we similarly get all the deferred inputs and we recursively
create another LazyEvaluator. Any deferred inputs will also be deferred
inside of this evaluator until we make an allocation.

Note that when any of the instructions has the same tensor as an input at two
different locations we force an allocation of the tensor.

If a deferred tensor is used by any other instructions which actually
requires the tensor (for example to do calculations), then the TensorMap
functions for getting input/output tensors will automatically create allocate
the tensor and call the post processing function (note that in this scenario
the tensor is most likely to be linearly allocate which might not be
optimal).

After the visitor has finished, any deferred allocations are propagated up
the scope to the callsite. Depending on whether the particular operand was
inplace at that callsite we might need to add a clone of the tensor.

Removed the old subcomputation and deferred allocation visitor and all the helper structs which were used by them.

Fixes T13355

Depends on D20595

Test Plan:
All tests pass.

Resnet50 fits when reducing the available memory proportion

Reviewers: #tensorflow, davidn

Reviewed By: #tensorflow, davidn

Subscribers: davidn

Maniphest Tasks: T13355

Differential Revision: https://phabricator.sourcevertex.net/D20838