poplar_executor: Abort when consuming past end of infeed iterator
Summary:
Before, when trying to consume more elements than were provided
by the infeed dataset iterator, we would block indefinetely
on SPSCQueue::BlockPop() in the Poplar stream callback.

We would also signal an out of range error in the normal case
of reaching the end of an infeed dataset iterator (without
attempting to go past it).

This changes the behaviour to aborting with a fatal error in
the case of consuming past the end of the iterator, and only
logging with less verbosity when reaching the end.

This is done by wrapping the SPSCQueue with support for pushing
a sentinel "end of queue" value. The producer pushes this value
when the infeed dataset iterator has reached the end. If the
consumer pops this value, it means that it has consumed past
the end, and we abort with an error.

Since this is in the time critical Poplar stream callback,
these functions are implemented inline in the header.

Based on feedback in T15954.

Test Plan:
All tests pass. Also ran some IPUEstimator tests on hardware.
Verified manually that program is aborted if we consume more
elements than we have.

Reviewers: #tensorflow, davidn, georgep

Reviewed By: #tensorflow, davidn, georgep

Subscribers: davidn, georgep

Differential Revision: https://phabricator.sourcevertex.net/D21118