Make outfeed queue compatible with TF2
Summary:
In TF2, creating operations in the initializer causes them to execute, so instead
create the operations on demand.  Also, TF2 tracing will create multiple graphs and
call the trace function mutliple times.  So we need to allow an outfeed enqueue
operation to be created multiple times, as long as they are in different graphs.

Test Plan: All tests passing.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T13604

Differential Revision: https://phabricator.sourcevertex.net/D21885