sendrecv_ops: Add support for concatenating output from replicas
Summary:
The previous behaviour was to always select the output from the
first replica. This was OK when doing replicated all-reduce of
the gradients before sending them. But when we instead do a
replicated reduce-scatter and send that we need to concatenate
the output from all replicas. Add an option for doing this.

T12685

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T12685

Differential Revision: https://phabricator.sourcevertex.net/D22216