Add RecvFromHost op with support for explicit in-placing
Summary:
This op takes an optional input that is used as the stream copy
destination if the op is in-place. If the op is not in-place,
it allocates a new tensor based on the layout desired by the
consumer, rather than copying the input tensor.

Lower to this from XlaHostCompute rather than Recv/RecvDone.
Use a simple heuristic that attempts to do it in-place if
there is an input tensor and output tensor from the outside
compilation scope with the same index and the same shape.

This heuristic works for the host-based reduction done in
`optimizer.apply_gradients()`. This decreases the memory
consumption compared to when poplar attempts to do the
aliasing itself, at least when testing ResNet-50.

Fixes T17726.

Test Plan:
Existing tests pass. Added/updated some tests.
Memory consumption for distributed ResNet-50 goes down.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T17726

Differential Revision: https://phabricator.sourcevertex.net/D23144