poplar_executor: Avoid copy of send stream buffer if deemed safe
Summary:
If the Poplar stream buffers cannot overlap and we are using the IPU
target which pre-allocates the host stream buffers, we can create a
non-owning tensor buffer view from the buffer pointer that we can use for
the `tensorflow::Tensor`.

This is safe since the lifetime of the buffer is longer than the lifetime
of the Tensor (the lifetime of the Poplar host buffer is until another
engine is loaded). Add some sanity checks for this assumption. It also
relies on the assumption that this same stream will not send again until
the data has been consumed, which we know for these send streams by design
(i.e. they cannot be used inside loops).

We also have to maintain a refcount to the TensorBuffer such that no
TensorFlow ops will attempt to write to it in-place.

It might seem like a set of fragile assumptions, but I think the sanity
checks mitigates the risk somewhat and the result is a quite good
performance boost (+7%) for distributed ResNet-50 training.

Fixes T18547.

Test Plan:
Since this is only used on hardware, we rely on the distributed training
test in the SDK tests (public_examples) for automated coverage of this.
I have also been running the ipu_multi_worker_strategy tests manually on
real hardware. Maybe I should copy some of them over to the
hardware_integration_tests repo.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T18547

Differential Revision: https://phabricator.sourcevertex.net/D23861